feat: Implement Foxhound connection
This is just me familiarising myself with Dart primarily, don't mind me.

Nevertheless, this could form a framework around which to organise
connection to Foxhound, at least in the data sense. (Notably, this
doesn't yet presuppose about where the connection would be instantiated
and/or rotated if the endpoint were not to be determined dynamically.
Neither does this actually do anything meaningful, given that I don't
have a means to test it yet. 😇️)