chore: Improve response de-/serialization
For the most part this is simply an elaboration of `fromJson` factories
as to not use a stupid single-purpose object that doesn't feel
particularly Dart-y. This has a couple of secondary implications though.

First, there's a new `Unjsoner` helper that makes sure that the incoming
data is well-formed and bails out early if that isn't the case.

In that light, this removes the `unknown` options from enums for both
indexing state and file type; those are validated eagerly as soon as
they come through from the API. This might be slightly too pedantic and
perhaps should be rolled back pending discussion.

The error construction was simplified to use the JSON-RPC native
representation which is already parsed and neatly categorised. This also
allows handling the token rotation case when needed (this handling is
forthcoming.)

Finally, some detritus from the previous boilerplate attempt was removed
where it had already been replaced by an actual implementation, as well
as some scaffolding moved around to better fit Dart conventions.