feat: Implement uploading
Hey ho, this is untested so no guarantees about how well it works.
Regardless, this should technically do uploads n stuff.

This retroopts the PifsClientProvider into providing an implementation
of a PifsUploader too. I get that this is kind of stretching the meaning
of "PIFS client provider" but, frankly, I don't see either a case where
the client should be real and the uploader fake (or vice versa), and I
don't want to bloat the widget tree with another provider. So this in
general should be fine.